Move helpers `getAbsolutePath`, `getNearestCommonDirectory` and `debugFilesPlugins` to `@dd/core`
