+++ packages/plugins/build-report/src/helpers.ts
@@ -5,7 +5,6 @@
 import { INJECTED_FILE } from '@dd/core/constants';
 import { isInjectionFile } from '@dd/core/helpers';
 import type { GlobalContext } from '@dd/core/types';
-import path from 'path';
 
 // Will match any last part of a path after a dot or slash and is a word character.
 const EXTENSION_RX = /\.(?!.*(?:\.|\/|\\))(\w{1,})/g;
@@ -80,18 +79,6 @@ export const cleanPath = (filepath: string) => {
             // sometimes added in rollup by the commonjs plugin.
             .replace(/^[^\w\s.,!@#$%^&*()=+~`\-/]+/, '')
     );
-};
-
-// Will only prepend the cwd if not already there.
-export const getAbsolutePath = (cwd: string, filepath: string) => {
-    if (isInjectionFile(filepath)) {
-        return INJECTED_FILE;
-    }
-
-    if (filepath.startsWith(cwd) || path.isAbsolute(filepath)) {
-        return filepath;
-    }
-    return path.resolve(cwd, filepath);
 };
 
 // Extract a name from a path based on the context (out dir and cwd).
