+++ packages/plugins/injection/src/helpers.ts
@@ -2,10 +2,9 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { doRequest, truncateString } from '@dd/core/helpers';
+import { doRequest, getAbsolutePath, truncateString } from '@dd/core/helpers';
 import type { Logger, ToInjectItem } from '@dd/core/types';
 import { InjectPosition } from '@dd/core/types';
-import { getAbsolutePath } from '@dd/internal-build-report-plugin/helpers';
 import { readFile } from 'fs/promises';
 
 import { AFTER_INJECTION, BEFORE_INJECTION, DISTANT_FILE_RX } from './constants';
