+++ packages/tools/src/helpers.ts
@@ -3,16 +3,8 @@
 // Copyright 2019-Present Datadog, Inc.
 
 import { ALL_BUNDLERS, SUPPORTED_BUNDLERS } from '@dd/core/constants';
-import { outputJsonSync, readJsonSync, serializeBuildReport } from '@dd/core/helpers';
-import type {
-    BundlerFullName,
-    BundlerName,
-    GetCustomPlugins,
-    GetPlugins,
-    GlobalContext,
-    IterableElement,
-    OptionsWithDefaults,
-} from '@dd/core/types';
+import { readJsonSync } from '@dd/core/helpers';
+import type { BundlerFullName, BundlerName, GetPlugins, OptionsWithDefaults } from '@dd/core/types';
 import { getContext } from '@dd/factory/helpers';
 import chalk from 'chalk';
 import { execFile, execFileSync } from 'child_process';
@@ -260,79 +252,3 @@ export const getBundlerPicture = (bundler: string) => {
 
 export const isInternalPluginWorkspace = (workspace: Workspace) =>
     workspace.name.startsWith('@dd/internal-');
-
-// Returns a customPlugin to output some debug files.
-type CustomPlugins = ReturnType<GetCustomPlugins>;
-export const debugFilesPlugins = (context: GlobalContext): CustomPlugins => {
-    const rollupPlugin: IterableElement<CustomPlugins>['rollup'] = {
-        writeBundle(options, bundle) {
-            outputJsonSync(
-                path.resolve(context.bundler.outDir, `output.${context.bundler.fullName}.json`),
-                bundle,
-            );
-        },
-    };
-
-    const xpackPlugin: IterableElement<CustomPlugins>['webpack'] &
-        IterableElement<CustomPlugins>['rspack'] = (compiler) => {
-        type Stats = Parameters<Parameters<typeof compiler.hooks.done.tap>[1]>[0];
-
-        compiler.hooks.done.tap('bundler-outputs', (stats: Stats) => {
-            const statsJson = stats.toJson({
-                all: false,
-                assets: true,
-                children: true,
-                chunks: true,
-                chunkGroupAuxiliary: true,
-                chunkGroupChildren: true,
-                chunkGroups: true,
-                chunkModules: true,
-                chunkRelations: true,
-                entrypoints: true,
-                errors: true,
-                ids: true,
-                modules: true,
-                nestedModules: true,
-                reasons: true,
-                relatedAssets: true,
-                warnings: true,
-            });
-            outputJsonSync(
-                path.resolve(context.bundler.outDir, `output.${context.bundler.fullName}.json`),
-                statsJson,
-            );
-        });
-    };
-
-    return [
-        {
-            name: 'build-report',
-            writeBundle() {
-                outputJsonSync(
-                    path.resolve(context.bundler.outDir, `report.${context.bundler.fullName}.json`),
-                    serializeBuildReport(context.build),
-                );
-            },
-        },
-        {
-            name: 'bundler-outputs',
-            esbuild: {
-                setup(build) {
-                    build.onEnd((result) => {
-                        outputJsonSync(
-                            path.resolve(
-                                context.bundler.outDir,
-                                `output.${context.bundler.fullName}.json`,
-                            ),
-                            result.metafile,
-                        );
-                    });
-                },
-            },
-            rspack: xpackPlugin,
-            rollup: rollupPlugin,
-            vite: rollupPlugin,
-            webpack: xpackPlugin,
-        },
-    ];
-};
